test(cmd): consolidate list tests into list_test.go
